+++ src/Telegram.php
@@ -391,14 +391,15 @@ class Telegram
     /**
      * Handle getUpdates method
      *
-     * @param int|null   $limit
-     * @param int|null   $timeout
-     * @param array|null $allowed_updates
+     * @todo Remove backwards compatibility for old signature and force $data to be an array.
+     *
+     * @param array|int|null $data
+     * @param int|null       $timeout
      *
      * @return ServerResponse
      * @throws TelegramException
      */
-    public function handleGetUpdates(?int $limit = null, ?int $timeout = null, ?array $allowed_updates = null): ServerResponse
+    public function handleGetUpdates($data = null, ?int $timeout = null): ServerResponse
     {
         if (empty($this->bot_username)) {
             throw new TelegramException('Bot Username is not defined!');
@@ -414,13 +415,31 @@ class Telegram
             );
         }
 
+        $offset          = 0;
+        $limit           = null;
+        $allowed_updates = null;
+
+        // @todo Backwards compatibility for old signature, remove in next version.
+        if (!is_array($data)) {
+            $limit = $data;
+
+            @trigger_error(
+                sprintf('Use of $limit and $timeout parameters in %s is deprecated. Use $data array instead.', __METHOD__),
+                E_USER_DEPRECATED
+            );
+        } else {
+            $offset          = $data['offset'] ?? $offset;
+            $limit           = $data['limit'] ?? $limit;
+            $timeout         = $data['timeout'] ?? $timeout;
+            $allowed_updates = $data['allowed_updates'] ?? $allowed_updates;
+        }
+
         // By default, allow ALL known update types.
         if ($allowed_updates === null) {
             $allowed_updates = Update::getUpdateTypes();
         }
-        $offset = 0;
 
-        //Take custom input into account.
+        // Take custom input into account.
         if ($custom_input = $this->getCustomInput()) {
             try {
                 $input = json_decode($this->input, true, 512, JSON_THROW_ON_ERROR);