+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Exclamation symbols (:exclamation:) note something of importance e.g. breaking c
 - Bot API 5.1 (ChatMember Update types, Improved Invite Links, Voice Chat). (@massadm, @noplanman)
 ### Changed
 ### Deprecated
+- `Telegram::handleGetUpdates` method should be passed a `$data` array for parameters.
 ### Removed
 ### Fixed
 - `message.edit_date` is now of type `timestamp`.